+++ Source Main 5.2/source/w_CharacterInfo.h
@@ -179,8 +179,8 @@ public:
 
     WORD        Level;
     EMonsterType        MonsterIndex;
-    WORD        Damage;
-    WORD        Hit;
+    int        Damage;
+    int        Hit;
     WORD        MoveSpeed;
 
     int			Action;
